More advanced simulator features: types and errors.
This commit adds sensor typing to the simulator. Each request must now come with
a type ID (and possibly the type's parameters) describing the values to be used
when simulating the sensors.

This also adds error handling through exceptions. Tokens are now handled, and
all requests must come with one. If an exception is raised by the payload, the
error will be sent on sensors/event, identified with a token. Successful
requests will also have feedback available there.