Adding announcements and moving MQTT code into the simulator.
This commit adds methods to handle the announcements channel (sensors), on which
the sensors collection is to be sent at regular intervals.

The MQTT-related code has been removed from its own mqtt.js file, and merged
directly into the simulator class. Having it separate is an inconvenient bit of
overkill.

Note: this commit cherry-picked the master branch .gitignore file.